new cat

+++ week3_greedy_algorithms/1_money_change/change.py
@@ -61,5 +61,5 @@ def get_change(m):
     change = sum(changed)
     return change
 
-m = float(input())
+m = int(input())
 print(get_change(m))
